+++ ec/interface.py
@@ -7,9 +7,19 @@ Allows the importing of ec, as a module.
 import shlex
 
 from modules import core
-from modules.core import execCommand, getDescendant
+from modules.core import execCommand, getDescendant, processPendingModules
+from modules import hooks
 
-__all__ = ['resolve', 'call']
+__all__ = ['setBase', 'resolve', 'call']
+
+# Exports
+def setBase(Underlying):
+  """Sets the base for the interface to work on.
+  """
+  from modules import state
+  
+  core.processPendingModules()
+  core.BaseGroup = Underlying.__ec_member__
 
 def resolve(route):
   """Resolves the member identified by the route.
@@ -32,4 +42,12 @@ def call(command, collect_missing=False):
     The return value of the called command.
   """
   return execCommand(shlex.split(command), collect_missing)
+  
+# main
+if hooks.isImportHooked(): # import hooks has been added by ec.FirstCaller, hence it wouldn't be processed automatically
+  processPendingModules()
+  
+else:
+  hooks.EcModuleName = '%s.ec' % __name__[:__name__.rfind('.')]
+  hooks.hookIntoImport()